Move Attributes into Parsing.Symbols
Attributes defines our representation of attributes and, thus, should be in the namespace for symbols.
