+++ RubberduckTests/Symbols/AccessibilityCheckTests.cs
@@ -48,7 +48,7 @@ namespace RubberduckTests.Symbols
             private static ClassModuleDeclaration GetTestClassModule(Declaration projectDeclatation, string name, bool isExposed = false)
             {
                 var qualifiedClassModuleMemberName = new QualifiedMemberName(StubQualifiedModuleName(), name);
-                var classModuleAttributes = new Rubberduck.Parsing.VBA.Attributes();
+                var classModuleAttributes = new Attributes();
                 if (isExposed)
                 {
                     classModuleAttributes.AddExposedClassAttribute();
